Fix corruption image is shown at external display when using Color Inversion
The point of problem is not to set mVpWidth and mVpHeight properly
according to display when color inversion is enabled.
It should be switched correctly before engine.beginGroup is called.
But it's done in doComposeSurfaces by the eglMakeCurrent.

So this patch changes the excution timing
to make sure the flow of engine.beginGroup/endGroup.

Change-Id: I30802e81496d4247f611d720b4a334def023770e